% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyCalibration.R
\name{applyCalibration}
\alias{applyCalibration}
\title{Apply Bearing Calibration}
\usage{
applyCalibration(stationList, myStations, recalibrate = FALSE, method = "",
  ...)
}
\arguments{
\item{stationList}{a list of sonobuoy stations as created by \code{loadStations},
or a single station as created by \code{formatStation}}

\item{myStations}{IDs of stations to check. Needed for calibrateStations to work correctly}

\item{recalibrate}{should buoys that have already been checked be re-examined? If
\code{FALSE}, any buoys with existing buoyQuality will be skipped over.}

\item{method}{method of angle calibration to use. \code{magnetic} uses
the median of magnetic varation values. \code{median} uses the median
of offset angles. \code{sine} uses sinusoidal fit to error.}

\item{\dots}{arguments passed on to other functions}
}
\description{
Return calibrated DIFAR angles using different methods
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
