% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyDrift.R
\name{applyDrift}
\alias{applyDrift}
\title{Apply Drift Calibration}
\usage{
applyDrift(stationList, myStations, recalibrate = FALSE, ...)
}
\arguments{
\item{stationList}{a list of sonobuoy stations as created by \code{loadStations},
or a single station as created by \code{formatStation}}

\item{myStations}{IDs of stations to check. Needed for calibrateStations to work correctly}

\item{recalibrate}{should buoys that have already been checked be re-examined? If
\code{FALSE}, any buoys with existing drift data will be skipped over.}

\item{\dots}{arguments passed on to other functions}
}
\description{
Applies estimated drift to detection and calibration data. Detection
  and calibration data frames will have two new columns \code{driftedBuoyLat}
  and \code{driftedBuoyLong}. These are calculated based on the drift rate, bearing,
  and time since deployment.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
