% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoyLoc.R
\name{buoyLoc}
\alias{buoyLoc}
\title{Buoy Location}
\usage{
buoyLoc(df)
}
\arguments{
\item{df}{data.frame of DIFAR data}
}
\description{
Get the median location of each buoy.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
