% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoyPosition.R
\name{buoyPosition}
\alias{buoyPosition}
\alias{buoyPositionInitial}
\title{Buoy Position}
\usage{
buoyPosition(x, buoy.data, method = c("initial", "interpolate", "drift"), ...)

buoyPositionInitial(x, buoy.data)
}
\arguments{
\item{x}{a data frame containing a column specifying buoy (\code{Buoy}) and
time (\code{UTC})}

\item{buoy.data}{data.frame of buoy positions. Format depends on choice of
\code{method}}

\item{method}{method of position estimation to use}
}
\description{
Return estimated position of buoys given a set of times
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
