% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOffset.R
\name{calculateOffset}
\alias{calculateOffset}
\title{Calculate Bearing Offset}
\usage{
calculateOffset(calibration, position, db)
}
\arguments{
\item{calibration}{calibration list of buoys for a station}

\item{position}{position list of buoys for a station}

\item{db}{database list of data frames for a station}
}
\description{
Calculate offset between DIFAR bearing and true bearing for
  calibration
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
