% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateStations.R
\name{calibrateStations}
\alias{calibrateStations}
\title{Calibrate Sonobuoy Stations}
\usage{
calibrateStations(stationList, type = "", recalibrate = FALSE, ...)
}
\arguments{
\item{stationList}{Either a list of stations or a single station to calibrate}

\item{type}{The type of calibration to perform, either \code{'check'}, \code{'apply'},
or \code{'drift'}.}

\item{recalibrate}{Should we re-do stations we have already calibrated?}

\item{\dots}{Other parameters to be passed to different calibration methods}
}
\value{
Returns the stationList object with calibration information added. Modifies
  the object in the parent environment so no assignment is needed.
}
\description{
Perform calibration for sonobuoy stations. This can take either a list
  of stations or a single station as input, and will modify the object directly in
  the parent environment so no assignment is necessary. The user can select from three
  calibration options: Check calibrations, apply calibrations, or calibrate drift (these
  must be done in this order).
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
