% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCalibrations.R
\name{checkCalibrations}
\alias{checkCalibrations}
\title{Check Buoy Calibrations}
\usage{
checkCalibrations(stationList, recheck = FALSE, goodRange = 10,
  midFun = median)
}
\arguments{
\item{stationList}{a list of sonobuoy stations as created by \code{loadStations},
or a single station as created by \code{formatStation}}

\item{recheck}{should buoys that have already been checked be re-examined? If
\code{FALSE}, any buoys with existing BuoyQuality will be skipped over.}
}
\value{
a list of sonobuoy stations with an added BuoyQuality field for each buoy.
  Will also directly modify the calStations object even if the result is not
  assigned to a variable.
}
\description{
Loop through histograms of buoy calibration offsets for each
  buoy in each station and mark the quality of the buoy as Good, Bad, or
  Questionable.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
