% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipBuoyLatLon.R
\name{clipBuoyLatLon}
\alias{clipBuoyLatLon}
\title{Clip data.frame for Buoy Location}
\usage{
clipBuoyLatLon(df, lat.range = NULL, lon.range = NULL)
}
\arguments{
\item{df}{data.frame of DIFAR data}

\item{lat.range}{latitude range to include}

\item{lon.range}{longitude range to include}
}
\description{
Remove buoys outside of requested range.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
