% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummaryReport.R
\name{createSummaryReport}
\alias{createSummaryReport}
\title{Create a Summary Report for a List of Stations}
\usage{
createSummaryReport(stationList, outDir = ".", fileName = "summaryReport",
  format = "word")
}
\arguments{
\item{stationList}{list of stations to create the summary report for}

\item{outDir}{directory to write the report to. Will create subfolders Report,
Report/Tables, and Report/Figures within this directory}

\item{fileName}{name of the report file}

\item{format}{output format of report, either word or html}
}
\description{
Creates a summary report to be used for cruise reports or as a starting
  point for publications. Saves the report as a word document, and saves all
  associated figures and tables as images in a Report folder. Automatically fills in
  summary information (number of sonobuoys deployed, detections, etc.), and prompts
  user for some information not contained in the databases. The created report should
  not be considered a final document, some information will need to be added / edited.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
