% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectionSummary.R
\name{detectionSummary}
\alias{detectionSummary}
\title{Create a Summary Dataframe of All Species Detections}
\usage{
detectionSummary(stationList)
}
\arguments{
\item{stationList}{list of sonobuoy stations returned by loadStations}
}
\description{
Return a dataframe containing the counts of all detected
   species at each station
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
