% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driftCalibration.R
\name{driftCalibration}
\alias{driftCalibration}
\title{Estimate Sonobuoy Drift Rate and Direction}
\usage{
driftCalibration(buoy.data, initial = c(1, 0), sd = 10, ...)
}
\arguments{
\item{buoy.data}{list of data.frames containing buoy calibration and
position data}

\item{initial}{vector of initial values for optim algorithm in km/h and true bearing}

\item{graph}{flag of whether or not to create likelihood graph for diagnostics}
}
\description{
Return estimated drift rate and direction of a sonobuoy
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
