% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effortPairs.R
\name{effortDuration}
\alias{effortDuration}
\title{Effort Duration}
\usage{
effortDuration(df)
}
\arguments{
\item{df}{data.frame of effort data}
}
\description{
Calculate duration of effort segments.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

