% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePosition.R
\name{estimatePosition}
\alias{estimatePosition}
\title{Estimate Position}
\usage{
estimatePosition(x, pos)
}
\arguments{
\item{x}{vector of times in \code{POSIXct} format}

\item{pos}{a data.frame with time in \code{POSIXct} in a column labelled
\code{UTC} and position in columns named \code{Latitude} and \code{Longitude}}
}
\value{
a data frame of estimated latitude and longitude for each element in
  \code{time}
}
\description{
Estimate positions for a set of times using a data frame of
  time and position combinations
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
