% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDifar.R
\name{formatDifar}
\alias{formatDifar}
\title{Format DIFAR data}
\usage{
formatDifar(df)
}
\arguments{
\item{df}{data.frame of DIFAR data from database}
}
\description{
Format DIFAR data from SQLite database.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
