% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatEffort.R
\name{formatEffort}
\alias{formatEffort}
\title{Format Effort}
\usage{
formatEffort(db)
}
\arguments{
\item{db}{list of data frames from PAMGuard SQLlite database}
}
\value{
a list of effort windows for each buoy
}
\description{
Format effort for each buoy by removing noise and creating
  blocks of alternating effort
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
