% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatStation.R
\name{formatStation}
\alias{formatStation}
\alias{formatBuoyPosition}
\alias{formatBuoyCalibration}
\alias{formatBuoyEffort}
\alias{formatDetections}
\alias{convertBinaryChannel}
\title{Format Station Data}
\usage{
formatStation(db, buoyPositions = NULL, overrideError = FALSE,
  dateFormat = "\%Y-\%m-\%d \%H:\%M:\%S", extraCols = NULL, ...)

formatBuoyPosition(db)

formatBuoyCalibration(db, position)

formatBuoyEffort(db)

formatDetections(db, buoys, extraCols = NULL)

convertBinaryChannel(x)
}
\arguments{
\item{db}{list of data frames from PAMGuard SQLlite database}

\item{buoyPositions}{optional argument, can be a data frame containing position
information for buoy deployments or a file path. Data frame must have columns \code{Buoy},
\code{UTC}, \code{Latitude}, and \code{Longitude}. If buoy positions are present
both in the database and this dataframe, positions from this data frame will be used.
File path must be to a csv file to be read in as a data frame.}

\item{overrideError}{station formatting will stop if insufficient buoy position data is
present. Set this to \code{TRUE} to override this behaviour and continue loading
the station and mark problematic buoys with \code{error=TRUE}.}

\item{dateFormat}{character string giving date-time format as used by strptime. This
is used to format the dates of data supplied by buoyPositions.}

\item{extraCols}{character vector listing any extra columns to keep while loading
detection data.}
}
\description{
Format data for a station from an list of tables read from
  a PAMGuard SQLlite database.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}, Taiki Sakai \email{taiki.sakai@noaa.gov}
}
\keyword{internal}
