% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMap.R
\name{getMap}
\alias{getMap}
\title{Get a Map of Appropriate Size}
\usage{
getMap(positions, force = FALSE, zoom = 10, center = NULL,
  quiet = FALSE)
}
\arguments{
\item{positions}{data frame containing columns \code{Latitude} and \code{Longitude}}

\item{force}{flag whether or not to force a specific zoom level instead of
finding it automatically}

\item{zoom}{zoom level used for the map. An integer, see zoom description in
get_map function of ggmap}

\item{center}{the location to center the map on. If \code{NULL}, will use the
mean of the range of the data. If not \code{NULL} it must be a named vector
with values \code{lon} and \code{lat}}

\item{quiet}{\code{FALSE} to report zoom level used}
}
\description{
Return a ggmap object centered on the middle of the positions
   data. Will start zoomed in then gradually zoom out until all points fit
   unless force is TRUE
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
