% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelDetection.R
\name{labelDetection}
\alias{labelDetection}
\title{Label Detections}
\usage{
labelDetection(df)
}
\arguments{
\item{df}{data.frame of DIFAR data}
}
\description{
Label common DIFAR detections.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
