% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDB.R
\name{loadDB}
\alias{loadDB}
\title{Load DIFAR database}
\usage{
loadDB(fname, verbose = FALSE)
}
\arguments{
\item{fname}{filename of SQLite database with DIFAR data. if \code{fname} is
missing a dialog box will be presented to choose a filename}

\item{verbose}{show database structure after loading?}
}
\value{
a list of data frames representing all tables in the supplied
  SQLite database. Date/time values in columns labelled \code{UTC} will be
  converted to \code{POSIXct}
}
\description{
Load list of data frames in DIFAR database
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
