% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDifar.R
\name{loadDifar}
\alias{loadDifar}
\title{Load DIFAR data}
\usage{
loadDifar(db.fname)
}
\arguments{
\item{db.fname}{filename of SQLite database with DIFAR data}
}
\description{
Load DIFAR data from SQLite database.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

