% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadEffort.R
\name{loadEffort}
\alias{loadEffort}
\title{Load effort data}
\usage{
loadEffort(folder)
}
\arguments{
\item{folder}{folder containing .csv files of effort}
}
\description{
Load effort data from folder.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
