% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNoise.R
\name{loadNoise}
\alias{loadNoise}
\title{Load noise data}
\usage{
loadNoise(folder)
}
\arguments{
\item{folder}{folder containing .csv files of noise}
}
\description{
Load noise data from folder.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

