% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDetections.R
\name{mapDetections}
\alias{mapDetections}
\title{Create a Map Showing Numbers of Detections}
\usage{
mapDetections(detectionData, bySpecies = "all", map = NULL, size = 3,
  nrow = 2, palette = "Reds")
}
\arguments{
\item{detectionData}{a dataframe with data on detections. Must have columns
\code{Latitude}, \code{Longitude}, \code{Station}, and \code{Count}. Must
also have column \code{Species} unless bySpecies is \code{'none'}}

\item{bySpecies}{The different species to group the counts by. If \code{'none'},
will group all detections together. If \code{'all'}, will show maps for all species.
Can also accept a vector of individual species to be graphed.}

\item{map}{Optional, a ggmap object to plot on. If left as NULL, will be created by the
getMap function. Can be included to reduce calls to getMap.}

\item{size}{Size of points to be plotted}

\item{nrow}{number of rows when plotting multiple species}

\item{palette}{color palette to be used, see brewer.pal for more info}
}
\description{
Return a ggmap object showing the number of detections at each sonobuoy
   station. Can show maps for all detections, or broken up by species.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
