% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapStations.R
\name{mapStations}
\alias{mapStations}
\title{Create a Summary Map of All Sonobuoy Stations}
\usage{
mapStations(detectionData, zoom = "auto", crop = FALSE, map = NULL,
  title = "Sonobuoy Stations", colorBy = "Cruise", size = 3)
}
\arguments{
\item{detectionData}{detection summary data returned by \code{detectionSummary}}

\item{zoom}{zoom level used for the map. Default is 'auto', will back out from
a zoom level of 10 until all data fits. If an integer, will force that zoom
level to be used.}

\item{crop}{flag whether or not to automatically crop the map to range of your data}

\item{map}{Optional, a ggmap object to plot on. If left as NULL, will be created automatically
by the getMap function. Can be included to reduce calls to getMap.}

\item{title}{a title for the plot}

\item{colorBy}{column to color points by}

\item{size}{size of points to plot}
}
\description{
Return a ggmap object showing all sonobuoy stations.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
