% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStationFile.R
\name{readStationFile}
\alias{readStationFile}
\title{Read station file}
\usage{
readStationFile(f)
}
\arguments{
\item{f}{.csv file of station data}
}
\description{
Read a file with station data into data.frame.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
