% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyCalibration.R
\name{applyCalibration}
\alias{applyCalibration}
\title{Apply Bearing Calibration}
\usage{
applyCalibration(stationList, myStations, recalibrate = FALSE, method = "",
  ...)
}
\arguments{
\item{method}{method of angle calibration to use. \code{magnetic} uses
the median of magnetic varation values. \code{median} uses the median
of offset angles. \code{sine} uses sinusoidal fit to error.}

\item{\dots}{arguments passed on to other functions}

\item{angle.data}{data.frame containing a \code{Buoy} column and a
\code{DIFARBearing} column with uncalibrated angle data}

\item{buoy.data}{list of data.frames containing buoy calibration data}
}
\description{
Return calibrated DIFAR angles using different methods
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
