% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCalibrations.R
\name{checkCalibrations}
\alias{checkCalibrations}
\title{Check Buoy Calibrations}
\usage{
checkCalibrations(stationList, myStations, recalibrate = FALSE,
  midFun = median, goodRange = 10)
}
\arguments{
\item{stationList}{a list of sonobuoy stations as created by \code{loadStations},
or a single station as created by \code{formatStation}}

\item{myStations}{IDs of stations to check. Neede for calibrateStations to work correctly}

\item{recalibrate}{should buoys that have already been checked be re-examined? If
\code{FALSE}, any buoys with existing BuoyQuality will be skipped over.}

\item{midFun}{function to use to show center of the data}

\item{goodRange}{range around center of data to show guidelines for good calibrations}
}
\value{
a list of sonobuoy stations with an added BuoyQuality field for each buoy.
}
\description{
Loop through histograms of buoy calibration offsets for each
  buoy in each station and mark the quality of the buoy as Good, Bad, or
  Questionable.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
