% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driftCalibration.R
\name{driftCalibration}
\alias{driftCalibration}
\title{Estimate Sonobuoy Drift Rate and Direction}
\usage{
driftCalibration(stationList, myStations, recalibrate = FALSE, map = FALSE)
}
\arguments{
\item{stationList}{a list of sonobuoy stations as created by \code{loadStations},
or a single station as created by \code{formatStation}}

\item{myStations}{IDs of stations to check. Needed for calibrateStations to work correctly}

\item{recalibrate}{should buoys that have already been checked be re-examined? If
\code{FALSE}, any buoys with existing buoyQuality will be skipped over.}

\item{map}{should we draw a map with ship position and estimated drift direction?}
}
\description{
Return estimated drift rate and direction of a sonobuoy
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
