% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatStation.R
\name{formatStation}
\alias{formatStation}
\alias{formatBuoyPosition}
\alias{formatBuoyCalibration}
\alias{formatBuoyEffort}
\alias{formatDetections}
\alias{convertBinaryChannel}
\title{Format Station Data}
\usage{
formatStation(db, buoyPositions = NULL, overrideError = FALSE, ...)

formatBuoyPosition(db)

formatBuoyCalibration(db, position)

formatBuoyEffort(db)

formatDetections(db, buoys, extraCols = NULL)

convertBinaryChannel(x)
}
\arguments{
\item{db}{list of data frames from PAMGuard SQLlite database}

\item{buoyPositions}{optional argument, a data frame containing position
information for buoy deployments. Data frame must have columns \code{Buoy},
\code{UTC}, \code{Latitude}, and \code{Longitude}. If buoy positions are present
both in the database and this dataframe, positions from this data frame will be used.}

\item{overrideError}{station formatting will stop if insufficient buoy position data is
present. Set this to \code{TRUE} to override this behaviour and continue loading
the station and mark problematic buoys with \code{error=TRUE}.}
}
\description{
Format data for a station from an list of tables read from
  a PAMGuard SQLlite database.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}, Taiki Sakai \email{taiki.sakai@noaa.gov}
}
\keyword{internal}
