% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatStation.R
\name{formatStation}
\alias{formatStation}
\alias{formatBuoyPosition}
\alias{formatBuoyCalibration}
\alias{formatBuoyEffort}
\alias{formatDetections}
\title{Format Station Data}
\usage{
formatStation(db)

formatBuoyPosition(db)

formatBuoyCalibration(db, position)

formatBuoyEffort(db)

formatDetections(db, buoys)
}
\arguments{
\item{db}{list of data frames from PAMGuard SQLlite database}
}
\description{
Format data for a station from an list of tables read from
  a PAMGuard SQLlite database.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
