% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDB.R
\name{loadDB}
\alias{loadDB}
\title{Load DIFAR database}
\usage{
loadDB(db.fname = NULL, db.str = FALSE)
}
\arguments{
\item{db.fname}{filename of SQLite database with DIFAR data}
}
\description{
Load list of data frames in DIFAR database
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
