% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDB.R
\name{loadDB}
\alias{loadDB}
\title{Load DIFAR database}
\usage{
loadDB(fname = NULL, show.str = FALSE)
}
\arguments{
\item{fname}{filename of SQLite database with DIFAR data}

\item{show.str}{show database structure?}
}
\description{
Load list of data frames in DIFAR database
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
