% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadStations.R
\name{loadStations}
\alias{loadStations}
\title{Load Multiple Stations}
\usage{
loadStations(folder, db.ext = "sqlite3", ...)
}
\arguments{
\item{folder}{folder containing SQLite databases with DIFAR data. if missing
a dialog box will be presented to choose a folder}
}
\value{
a list of sonobuoy stations
}
\description{
Load a list of sonobuoy stations from SQLlite databases in a folder
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
