% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapStations.R
\name{mapStations}
\alias{mapStations}
\title{Create a Summary Map of All Sonobuoy Stations}
\usage{
mapStations(stationList, zoom = "auto", crop = FALSE, map = NULL)
}
\arguments{
\item{stationList}{list of sonobuoy stations returned by loadStations}

\item{zoom}{zoom level used for the map. Default is 'auto', will back out from
a zoom level of 10 until all data fits. If an integer, will force that zoom
level to be used.}

\item{crop}{flag whether or not to automatically crop the map to range of your data}

\item{map}{Optional, a ggmap object to plot on. If left as NULL, will be created automatically
by the getMap function. Can be included to reduce calls to getMap.}
}
\description{
Return a ggmap object showing all sonobuoy stations.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
